rename typeProvider.d.ts to type-provider.d.ts
